fix for wrong dimensions
There was a bug in which my 2nd monitor, with different scaling option (125%), would not have its background generated with the correct dimensions.
Querying the device's scale factor yielded incorrect results (always reported 100%).
Turns out that this issue has something to do with the program & how it detects DPI settings. By setting the application mode to be per-monitor, the screen resolutions are now reported correctly.
In addition to setting the application mode, I also had to change forms' `AutoScaleMode` so they behave correctly depending on the noitor they're being display on.